refactor: config security refactor
