+++ packages/config/src/agent.ts
@@ -1,0 +1,12 @@
+// FIXME: this is fixed, should pick the package.json or official version
+import assert from 'assert';
+import _ from 'lodash';
+
+const pkgVersion = require('../package.json').version;
+const pkgName = require('../package.json').name;
+
+export function getUserAgent(): string {
+  assert(_.isString(pkgName));
+  assert(_.isString(pkgVersion));
+  return `${pkgName}/${pkgVersion}`;
+}
