+++ packages/config/src/conf/docker.yaml
@@ -86,12 +86,11 @@ logs:
   # FIXME: this should be documented
 # More info about log rotation https://github.com/pinojs/pino/blob/master/docs/help.md#log-rotation
 
-#experiments:
-#  # support for npm token command
-#  token: false
-#  # support for the new v1 search endpoint, functional by incomplete read more on ticket 1732
-#  search: false
-
+flags:
+  # support for npm token command
+  token: false
+  # support for the new v1 search endpoint, functional by incomplete read more on ticket 1732
+  search: false
 # This affect the web and api (not developed yet)
 #i18n:
 #web: en-US
