--- packages/config/src/conf/full.yaml
@@ -1,10 +1,0 @@
-## This file is deprecated and the content does not exist anymore
-## we highly recommend either visit
-## https://verdaccio.org/docs/en/configuration
-## or read the local file
-## https://github.com/verdaccio/website/tree/master/docs/config.md
-
-## contribute with translations
-
-## You can contribute translating documentation through the crowdin platform
-## https://crowdin.com/project/verdaccio
