+++ packages/config/src/index.ts
@@ -1,5 +1,7 @@
 export * from './config';
 export * from './config-path';
 export * from './token';
-export * from './config-utils';
+export * from './package-access';
 export * from './parse';
+export * from './uplinks';
+export * from './security';
