+++ packages/config/src/string.ts
