+++ packages/config/test/package-access.spec.ts
@@ -3,55 +3,16 @@ import _ from 'lodash';
 
 import { PACKAGE_ACCESS } from '@verdaccio/dev-commons';
 
-import { spliceURL } from '@verdaccio/utils';
-import {
-  getMatchedPackagesSpec,
-  hasProxyTo,
-  normalisePackageAccess,
-  sanityCheckUplinksProps,
-  uplinkSanityCheck,
-} from '../src/config-utils';
+import { getMatchedPackagesSpec, normalisePackageAccess } from '../src/package-access';
 import { parseConfigFile } from '../src';
 
-describe('Config Utilities', () => {
+describe('Package access utilities', () => {
   const parseConfigurationFile = (conf) => {
     const { name, ext } = path.parse(conf);
     const format = ext.startsWith('.') ? ext.substring(1) : 'yaml';
 
     return path.join(__dirname, `./partials/config/${format}/${name}.${format}`);
   };
-
-  describe('uplinkSanityCheck', () => {
-    test('should test basic conversion', () => {
-      const uplinks = uplinkSanityCheck(
-        parseConfigFile(parseConfigurationFile('uplink-basic')).uplinks
-      );
-      expect(Object.keys(uplinks)).toContain('server1');
-      expect(Object.keys(uplinks)).toContain('server2');
-    });
-
-    test('should throw error on blacklisted uplink name', () => {
-      const { uplinks } = parseConfigFile(parseConfigurationFile('uplink-wrong'));
-
-      expect(() => {
-        uplinkSanityCheck(uplinks);
-      }).toThrow('CONFIG: reserved uplink name: anonymous');
-    });
-  });
-
-  describe('sanityCheckUplinksProps', () => {
-    test('should fails if url prop is missing', () => {
-      const { uplinks } = parseConfigFile(parseConfigurationFile('uplink-wrong'));
-      expect(() => {
-        sanityCheckUplinksProps(uplinks);
-      }).toThrow('CONFIG: no url for uplink: none-url');
-    });
-
-    test('should bypass an empty uplink list', () => {
-      // @ts-ignore
-      expect(sanityCheckUplinksProps([])).toHaveLength(0);
-    });
-  });
 
   describe('normalisePackageAccess', () => {
     test('should test basic conversion', () => {
@@ -205,115 +166,6 @@ describe('Config Utilities', () => {
       expect(getMatchedPackagesSpec('@fake/angular', packages).proxy).toMatch('npmjs');
       expect(getMatchedPackagesSpec('vue', packages)).toBeUndefined();
       expect(getMatchedPackagesSpec('@scope/vue', packages)).toBeUndefined();
-    });
-  });
-
-  describe('hasProxyTo', () => {
-    test('should test basic config', () => {
-      const packages = normalisePackageAccess(
-        parseConfigFile(parseConfigurationFile('pkgs-basic')).packages
-      );
-      // react
-      expect(hasProxyTo('react', 'facebook', packages)).toBeFalsy();
-      expect(hasProxyTo('react', 'google', packages)).toBeFalsy();
-      // vue
-      expect(hasProxyTo('vue', 'google', packages)).toBeFalsy();
-      expect(hasProxyTo('vue', 'fake', packages)).toBeFalsy();
-      expect(hasProxyTo('vue', 'npmjs', packages)).toBeTruthy();
-      // angular
-      expect(hasProxyTo('angular', 'google', packages)).toBeFalsy();
-      expect(hasProxyTo('angular', 'facebook', packages)).toBeFalsy();
-      expect(hasProxyTo('angular', 'npmjs', packages)).toBeTruthy();
-    });
-
-    test('should test resolve based on custom package access', () => {
-      const packages = normalisePackageAccess(
-        parseConfigFile(parseConfigurationFile('pkgs-custom')).packages
-      );
-      // react
-      expect(hasProxyTo('react', 'facebook', packages)).toBeTruthy();
-      expect(hasProxyTo('react', 'google', packages)).toBeFalsy();
-      // vue
-      expect(hasProxyTo('vue', 'google', packages)).toBeFalsy();
-      expect(hasProxyTo('vue', 'fake', packages)).toBeFalsy();
-      expect(hasProxyTo('vue', 'npmjs', packages)).toBeTruthy();
-      // angular
-      expect(hasProxyTo('angular', 'google', packages)).toBeTruthy();
-      expect(hasProxyTo('angular', 'facebook', packages)).toBeFalsy();
-      expect(hasProxyTo('angular', 'npmjs', packages)).toBeFalsy();
-    });
-
-    test('should not resolve any proxy', () => {
-      const packages = normalisePackageAccess(
-        parseConfigFile(parseConfigurationFile('pkgs-empty')).packages
-      );
-      // react
-      expect(hasProxyTo('react', 'npmjs', packages)).toBeFalsy();
-      expect(hasProxyTo('react', 'npmjs', packages)).toBeFalsy();
-      // vue
-      expect(hasProxyTo('vue', 'npmjs', packages)).toBeFalsy();
-      expect(hasProxyTo('vue', 'npmjs', packages)).toBeFalsy();
-      expect(hasProxyTo('vue', 'npmjs', packages)).toBeFalsy();
-      // angular
-      expect(hasProxyTo('angular', 'npmjs', packages)).toBeFalsy();
-      expect(hasProxyTo('angular', 'npmjs', packages)).toBeFalsy();
-      expect(hasProxyTo('angular', 'npmjs', packages)).toBeFalsy();
-      // private
-      expect(hasProxyTo('private', 'fake', packages)).toBeFalsy();
-    });
-  });
-
-  describe('spliceURL', () => {
-    test('should splice two strings and generate a url', () => {
-      const url: string = spliceURL('http://domain.com', '/-/static/logo.png');
-
-      expect(url).toMatch('http://domain.com/-/static/logo.png');
-    });
-
-    test('should splice a empty strings and generate a url', () => {
-      const url: string = spliceURL('', '/-/static/logo.png');
-
-      expect(url).toMatch('/-/static/logo.png');
-    });
-  });
-
-  describe('JSON', () => {
-    test('parse default.json', () => {
-      const config = parseConfigFile(parseConfigurationFile('default.json'));
-
-      expect(config.storage).toBeDefined();
-    });
-
-    test('parse invalid.json', () => {
-      expect(function () {
-        parseConfigFile(parseConfigurationFile('invalid.json'));
-      }).toThrow(/Error/);
-    });
-
-    test('parse not-exists.json', () => {
-      expect(function () {
-        parseConfigFile(parseConfigurationFile('not-exists.json'));
-      }).toThrow(/Error/);
-    });
-  });
-
-  describe('JavaScript', () => {
-    test('parse default.js', () => {
-      const config = parseConfigFile(parseConfigurationFile('default.js'));
-
-      expect(config.storage).toBeDefined();
-    });
-
-    test('parse invalid.js', () => {
-      expect(function () {
-        parseConfigFile(parseConfigurationFile('invalid.js'));
-      }).toThrow(/Error/);
-    });
-
-    test('parse not-exists.js', () => {
-      expect(function () {
-        parseConfigFile(parseConfigurationFile('not-exists.js'));
-      }).toThrow(/Error/);
     });
   });
 });
