+++ packages/node-api/package.json
@@ -29,6 +29,7 @@
     "@verdaccio/config": "workspace:5.0.0-alpha.0",
     "@verdaccio/utils": "workspace:5.0.0-alpha.0",
     "lodash": "^4.17.20",
+    "debug": "^4.2.0",
     "core-js": "^3.6.5",
     "selfsigned": "1.10.7"
   },
