+++ packages/utils/src/index.ts
@@ -1,5 +1,4 @@
 export * from './auth-utils';
-export * from './string';
 export * from './utils';
 export * from './crypto-utils';
 export * from './replace-lodash';
